feat:push_noti

+++ lib/data/api/map/response/create_give_ride_response/create_give_ride_data_response.dart
@@ -16,7 +16,7 @@ class CreateGiveRideDataResponse {
   String? startTime;
   @JsonKey(name: 'end_time')
   String? endTime;
-  int? fare;
+  double? fare;
   CreateGiveRideVehicleResponse? vehicle;
 
   CreateGiveRideDataResponse({
