+++ lib/domain/map/output/create_give_ride/create_give_ride_output.dart
@@ -11,7 +11,7 @@ class CreateGiveRideOutput {
   final String? polyline;
   final int? distance;
   final int? duration;
-  final int? fare;
+  final double? fare;
   final String? vehicle;
   final String? vehicleId;
